Move stores to a separate module
